finished routes and components
