+++ src/components/AccountForm/AccountDetails/EditAccountEmail/index.js
@@ -1,0 +1,2 @@
+import EditAccountEmail from './EditAccountEmail';
+export default EditAccountEmail;
