+++ src/components/AccountForm/AccountDetails/EditAccountPassword/index.js
@@ -1,0 +1,2 @@
+import EditAccountPassword from './EditAccountPassword';
+export default EditAccountPassword;
