+++ src/components/AccountForm/AccountDetails/index.js
@@ -1,0 +1,2 @@
+import AccountDetailsHome from './AccountDetails';
+export default AccountDetailsHome;
