+++ src/components/AccountForm/AccountForm.js
@@ -5,6 +5,7 @@ import {
 } from 'react-router-dom'; 
 
 import AccountCreate from './AccountCreate';
+import AccountDetailsHome from './AccountDetails/AccountDetailsHome.js';
 import AccountLogin from './AccountLogin';
 
 function AccountForm() {
@@ -16,6 +17,10 @@ function AccountForm() {
 
       <Route path="/account/login">
         <AccountLogin />
+      </Route>
+
+      <Route path="/account/details">
+        <AccountDetailsHome />
       </Route>
     </Switch>
   );
